feat(server): 完成日志功能
- 添加日志功能
- 修复命令行参数的bug

+++ server/include/server.h
@@ -1,11 +1,13 @@
 #ifndef SERVER_H
 #define SERVER_H
 
+#include "define.h"
+
 #define LISTEN_PORT_LENGTH 5
 #define DEFAULT_LISTEN_PORT 8000
 #define BUFFER_LENGTH 1024
 
-extern int listen_port;//监听端口号
+extern u16 listen_port;//监听端口号
 extern int address_length;
 extern char send_buffer[BUFFER_LENGTH];
